Phase 4.1: Split repo_structure_lib.py into focused modules
repo_structure_lib.py was a ~470-line grab-bag mixing dataclasses, type
aliases, path helpers, capture/substitute utilities, exceptions and
scanning helpers. Split it along those seams:

- models.py    RepoEntry, Entry, Flags, ScanIssue, MatchResult and the
               type aliases plus the builtin rule constants
- paths.py     normalize_path, join_path_normalized, rel_dir_to_map_dir,
               map_dir_to_rel_dir
- patterns.py  substitute_pattern_captures, extract_pattern_captures,
               has_template_substitution, expand_companion_requirements
- errors.py    the exception classes
- scanning.py  skip_entry, to_entry, get_matching_item_index,
               check_companion_files and the backlog builders

repo_structure_lib stays as a thin re-export shim for one release so
existing imports keep working; all in-tree modules now import from the
new homes. _has_template_substitution becomes public, since it is now
used across the patterns/scanning boundary.

The unit tests move alongside their modules (paths_test.py,
patterns_test.py, scanning_test.py) with new coverage for the plain data
types and exceptions; repo_structure_lib_test.py now only asserts the
shim re-exports the same objects.

Closes #180
